Log - 1 - Animations/Controls
Added a few animations tying them up with the current controller that I made yesterday. As of now there are two animations, one for idle position of the player model and one for walking. Other than that, yesterday when I started the actual project i set the scene, made some adjustments to the camera settings and volume, something of note as I never had thought about this in any previous projects.